Patch 1.1
In this patch we addressed many bug fixes and added many new features!

List of changes and new stuff:

- Fixed the border issues with Flex Boxes
- Added Propper Font for the text
- Fixed the issue that the Border Weight didn't apply properly
To  Logo text by replacing it with an Image.
- Added Images!!! We finally have a little bit of a flair up in here
(And they're margenlized correctly!!!)
and it's a (lot) more bearable to look at now!
- Added an extra layer of visibility to the newly added BIG title text,
by reducing the background opacity, adding drop shadow and a 2nd layer
slightly offseted to the original!
- Added a (non-functional) Button! But hey, it's there!
- Fixed an issue where fonts didn't appear properly."

Dev Notes with Ben:

This patch was probably the most challenging part so far
but my knowledge of both html and CSS have exponentially been
Increased.

That's all I got to report for now

See you next patch!!